Improve support for DTRACE_DEBUG while testing.
You have long been able to set DTRACE_DEBUG in the environment of the
testsuite to get debugging output from a dtrace under test, but the
results are somewhat unsatisfactory: many tests fail due to
expected-results comparison failures, and some debugging output is not
provided because it was thrown away by test-result postprocessors.

Fix this by parsing the debugging output out of the stderr test results
before postprocessing, then dropping it into the sumfile after results
comparison.

Also eliminate debug-mode noise when running dtrace for non-test reasons
such as module autoloading, and tweak the dtrace debugging machinery so
that setting DTRACE_DEBUG to the empty string is the same as unsetting
it, so that you can turn off debugging when DTRACE_DEBUG might already
be set in the environment by running dtrace via
"DTRACE_DEBUG= dtrace ..." (which technically *sets* the environment
variable rather than unsetting it).