Revamp "crucial processes" system-daemon detection.
For some time now we have had code in DTrace which endeavours to determine if a
process is so crucial to the operation of the system that we shouldn't drop
breakpoints in it unless specifically commanded to do so via dtrace -p, so that
shortlived and otherwise relatively frequent operations for which quality of
name resolution is not utterly crucial (like ustack()) will not trigger a
ptrace() on things like your init process, possibly killing it if dtrace is kill
-9ed before it can remove its breakpoints.

Unfortunately the mechanism used for this was less than ideal (as I thought it
might be at the time I wrote it).  We decided that a process was a crucial
system daemon if its uid was zero and it didn't have a controlling terminal, but
this errs in both directions:

 - you can run the testsuite without a controlling terminal: tests then fail
   because parts of dtrace's own testsuite are considered crucial system
   daemons!

 - lots of crucial system daemons, like named, rpcbind and dbus-daemon run as
   non-root.

We employ a double-pronged approach to fix this, each of which adds a new
optional dtrace-compile-time option (which Oracle Linux users should rarely or
never need to use).

On systemd systems, like OL7, things are simple: if the process is in the
system.slice or the root slice, it is a crucial system daemon: everything else
(the user and machine slices and whatever other slices the sysadmin adds) is
fair game.  We can easily determine if a systemd system is in use as a one-off
operation when the first process is grabbed, because on any such system all
processes have a systemd cgroup line in their /proc/$self/cgroup: for all
processes not in the root slice, the slice name is also given there.  Of course,
the name of the system.slice is purely convention, so we have to add a new
"sysslice" option to notify dtrace if it's been renamed.

For non-systemd systems, like OL6, things are more heuristic.  We use the
existing algorithm as a starting point (root uid, no controlling terminal), but
rather than only looking at root-owned processes, we add a new configuration
option "useruid" that denotes the first UID that is not in the reserved system
range, and default it by grepping for UID_MIN in /etc/login.defs at build time.
(If that's not found, we use a default default of 1000, derived from the shadow
password suite's source code.)

But that's not enough -- it brings back the very bug we're trying to fix, that
the algorithm trips on the testsuite itself when run as root.  So we look at the
open file descriptors of such processes for their three standard streams: if any
of them are a TTY (even if not a controlling terminal) or a regular file, this
is not a system daemon but just a nohupped ordinary process being run in the
reserved range (this really happens, often when you're trying to debug
malfunctioning daemons, the very time when you want dtrace to catch them).  That
catches normal processes on terminals which aren't in their foreground process
group, and backgrounded processes, even if they have redirected output: it even
catches nohupped processes.

But that's not quite enough -- some processes satisfy all of these constraints
but are still system daemons.  One notable example is rsyslogd, which has open
standard streams to logfiles in /var/log, but things like databases (with the db
in the standard location under /var) and other things might reasonably match it
too.  So, as a final condition, if a process has an open fd to a regular file in
a standard stream *but that regular file is under /var* that fd is not
considered evidence that this is not a system daemon process.

Add a test for this too (which currently fails, because it also checks that
symbol resolution still sort of works in system daemons, and it is broken).

Orabug: 21914902