Make test timeouts an error.
This is only true by default: timeouts can be made non-fatal again via the
@@timeout-success option (which is turned on for test/demo, because these
originated as demo programs that are *meant* to time out), and via the
--ignore-timeouts option, which is meant for use when a stress test is
underway and the system is under load.

We also change the semantics of the @@timeout option: it now specifies
a minimum, so you cannot break a test with a long @@timeout when you
are trying to *boost* timeouts via, e.g. -- timeout=30.

This causes a number of tests to fail due to inadequate @@timeouts:
add appropriate @@timeout options to all of them.