Fix noninvasive tracing.
The process control thread in dtrace sits around and waits for incoming
ptrace() requests, proxy requests from the rest of dtrace, or process
termination and handles them.  Since noninvasive tracing does not
use ptrace(), it doesn't need a process control thread: so that thread
commits suicide as soon as it realises that this process will be
noninvasively traced.

This then runs the pthread cleanup hook for this thread, and
unfortunately this was not adjusted to allow for the possibility of
suiciding threads whose libproc handle was still useful.  So it destroys
the libproc handle, rendering lookups impossible.  It also falsely told
the rest of dtrace that the monitored process had terminated, when of
course we simply cannot determine that from userspace in the absence of
ptrace() (at least not without tiresome and expensive /proc scanning
which we are not doing for such a minor thing).

Fixing that reveals another problem: on dtrace shutdown (or rather
dtrace_close()), the destruction function for the hash table containing
dt_proc structures assumes that any dt_proc without a daemon thread is
entirely dead and needs only to be freed.  This is no longer true now
that daemon threads can suicide: do a preventative Prelease() just in
case.

Orabug: 22106441