Avoid %ifarch x86_64.
For reasons yet unclear, the arch on x86_64 for non-kernel builds on the
autobuilders is neither x86_64 nor x86 (maybe it's amd64 or something,
who knows).  This torpedoes the machinery we use to select the kernels
we need to build dtrace-utils against (specifically, the ones we need to
install devel packages for so we can substitute library .d scripts
against the #defines in their header files).

For sparc, we know that %ifarch sparc64 works: and elsewhere in the
specfile we do use %ifnarch sparc64 ... %else ... %endif constructs.
So migrate the dtrace_kernels machinery to the same construct.