tst.plddGrab*.sh: Do not try to pldd before the test program is exec()ed
We were seeing spurious, hard-to-reproduce failures from all the
plddGrab tests in which the process being traced seemed to have grown
all sorts of extra libraries: libtinfo and the like.

This turns out to be a race condition: we run the program via bash's &
operator, then immediately pldd it -- but the pldd races with bash's
exec() of the test program, and if it wins the race we end up pldding
bash itself! So we get a quite different set of shared libraries, and
a test failure.

The fix is to sit and busywait in a loop until the program is execed, as
revealed by a change in the value of /proc/$pid/exe away from bash.

Orabug: 21431540