Prevent debugging output from differnt threads from being intermingled.
It is possible for debugging output to be intermingled if two threads
call dt_debug() at the same time. This leads to output looking like
this:

libdtrace DEBUG 1447698316: libproc DEBUG 1447698316: blah
blah

which the testsuite then interprets as one line of debugging output and
one line of output on stderr, leading to spurious test failures.

The fundamental flaw is that while fprintf() et al lock the stream
around a single call, we emit the debugging output in multiple calls, so
we need to explicitly lock the stream around those calls to prevent
intermingling.