Make test --timeout specify a minimum.
This makes it safe to boost timeouts by specifying a --timeout greater
than the default of 10s.

Also fix a number of tests that specify inadequate @@timeouts and thus
were always failing.

This is a partial backport of the userspace-master commit "Make test
timeouts an error." Test timeouts are still not an error on this branch.