Unknown-type-lookup reduction: add more chars that cannot appear in types
We were already noting that identifiers in global context that that
contain colons or match 'BEGIN', 'END' or 'ERROR' could not be types and
thus should not trigger a lookup to see if they are -- but we can extend
this to everything in RGX_PSPEC that is not valid in a C identifier.
That would be, at least, the set [-.?!`] in addition to the hyphen we're
already checking.

This causes tests that look up "tick-1s" etc not to trigger loads of
all the CTF files.

Also added to the appropriate test to ensure this doesn't regress.