Read module name -> path mappings from the module.order file.
Rather than carrying out a filesystem walk to get the name -> path for
every kernel module in the system, read the module.order file to get the
list for all in-tree modules, then do a filesystem walk for the
remainder alone (i.e. for out-of-tree modules only).  If there is no
module.order file, or it is not readable, fall back to the old walk for
everything.

This is a micro-optimization until the disk is seriously loaded, when it
suddenly becomes much more important.  "Null" runs of dtrace via
dtrace -n 'BEGIN { exit(0); }' with a dropped cache between every dtrace
invocation suggest a reduction of startup time of approximately 50% and
a rough halving of I/O:

     baremetal (one disk): .233 user .073 system .851 total 123016 disk
                    after: .238 user .057 system .483 total 48648 disk

vm (4-way h/w RAID array): .459 user .186 system .927 total 182304 disk
                    after: .448 user .127 system .630 total 104360 disk

(The remaining I/O is accounted for by the small fs walk needed to dig
up out-of-tree modules such as dtrace.ko itself, and the >1MiB read
needed to pull in the modules.order file and the shared and vmlinux CTF
information.)

Testsuite runs with a similar dropped cache between every dtrace
invocation show a reduction in testsuite runtime on the order of 10% (57
mins -> 50 mins on the baremetal box; 50 mins -> 45 mins on the VM:
without dropped caches, the testsuite takes 29 mins to run on the
baremetal box, and 39 on the VM, not that that's very useful, but it
gives you an idea of the time hit of the CTF read and of loading dtrace
itself a couple of thousand times).

With hot caches, we see about a 1% reduction in runtime in all of these
tests, and never an increase.