Environment variable control of testsuite flags.
Many ./runtest.sh command line options can now be controlled via environment
variables, allowing you to set them when doing a 'make check'. They are still
overridden by command-line options, which means one or two don't work for 'make
check' but only with ./runtest.sh because 'make check' itself passes in the
corresponding command- line option.

The new variables, and their corresponding options, are:

Option                Variable
--timeout             DTRACE_TEST_TIMEOUT
--capture-expected    DTRACE_TEST_CAPTURE_EXPECTED
--use-installed       DTRACE_TEST_USE_INSTALLED (mostly redundant)
--testsuites=...      DTRACE_TEST_TESTSUITES (doesn't work with make check)
--baddof              DTRACE_TEST_BADDOF (probably a broken option)
--no-execute          DTRACE_TEST_NO_EXECUTE (probably useless)

As the list above shows, many are not much use or correspond to
long-untested options: the only really useful ones are likely to be
DTRACE_TEST_CAPTURE_EXPECTED and DTRACE_TEST_TIMEOUT.

(While we're at it, flip the sign of the internal baddof variable
around: it is almost always off, so we should be tracking NOBADDOF as
the positive case, not BADDOF.)