Revised modules-location code.
Reading modules.order doesn't work very well if modules install things
under kernel/, as the dtrace modules package does.  Instead, read
modules.dep: it's bigger, but that's not much of a problem if we boost
the libio buffer size for that file handle: the resulting code is
actually about 20% again faster than it was before, despite doing
33% more startup-time disk I/O.

We know modules.dep must always be present, complete, and readable for a
module to be usable at all, so we can rip out all the nftw()-related
code as well.

Orabug: 22237449
Orabug: 22238204