Fix tests relying on timeouts.
A number of tests in test/demo/ were relying on timeouts for successful
completion.  In addition to needlessly slowing testsuite execution, this
makes it impossible to boost the timeout so that it is (as originally
envisaged) a mechanism for halting *buggy* dtrace tests that do not
correctly halt.

This is a precondition for enabling timeout-as-error everywhere.