Make the dif bcopy test deterministic.
Right now, it is probing syscall::open:entry without actually ensuring
that anything is running that might open a file, so it takes a more or
less random time to run and might easily time out.

The pre-existing test/triggers/open.c is made for just this case, so
use it.