Fix extraction of test options from per-arch test.options files.
While exist_options() was adjusted for per-arch test.options files,
extract_options never was, so you could tell an option existed in
such a per-arch file but never get its value.  In conjunction with
the bug fixed in the previous commit this caused any attempt to
extract such an option to loop forever, giving the appearance of
a testsuite hang.

Orabug: 22030161