Change @@trigger to be synchronous by default.
Introduce a new @@trigger-timing value, 'synchro', which is now the
default.  This causes @@trigger to invoke the trigger by passing
-c and then the trigger name to dtrace: as such, the execution is
now synchronous.

Adjust tests to either work this way or to specify @@trigger-timing:
adjust the mmap trigger to no longer engage in useless synchronization
that is now done for it by dtrace.