Improve things when waitfd() fails completely.
Currently, if the first call to waitfd() fails (perhaps because you're
running a kernel that does not support it), the control thread dies
but most of dtrace does not notice, and proceeds to try to talk down
condition variables to a thread that no longer exists.  This has a
tendency to deadlock -- later calls might even infloop.

Back in the past (when this code was written) there was little that
could be done about this, but now that we have working noninvasive grabs
we can improve things a little by demoting a process whose control-
thread waitfd() fails to a noninvasive grab.  This means we can't get
good symbol resolution, etc, but at least we don't deadlock.