Copy data from high addresses in victim processes reliably.
In a fit of incompetence absolutely typical of Linux development, you
can read only *half* of victim memory from /proc/$pid/mem, because the
half above LONG_MAX is out of the range of a loff_t.  This is true even
though a loff_t is only signed for the sake of relative seeks, and
pread() is an absolute seeker!  Nobody cares about this ridiculous mess
because no real file is ever this large and x86 and ARM (the only
platforms anyone cares about) don't store anything useful that high up
in memory.  But SPARC does: all the shared libraries sit up there.

So we have to use PTRACE_PEEKDATA to pull stuff out of them in that
case.  Alignment constraints mean we can't just *read* it -- we have to
read in word-sized chunks, which means shaving the bottom bits off the
address, reading that much *more* into a temporary buffer, a few bytes
at a time, painfully slowly, then copying the appropriate bits out of
the middle of the buffer.