Fix Pget_bkpt_ip() return value.
This function returns an arbitrary pointer, so in order to be able to
represent addresses in the upper half of the address space it must
return a uintptr_t, not a long.  Two unlikely addresses remain
unrepresentable: zero (no breakpoint) and (uintptr_t)-1 (error).

All arch implementations already conform to this contract: all we need
to do is fix a few declarations.