Adjust tests for PID 0 psinfo.
The PID 0 task was not having psinfo populated prior to the fix for
bug 22561297, which caused builtinvar/tst.psinfo.d to output a lot
of memory access failures.  This commit limits that test to only
try to access psinfo for tasks with PID != 0.

This commit also adds a new test (builtinvar/X) that ensures that
we can access psinfo for PID 0.

Orabug: 22561297
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>