Remove dependency on a trigger for tst.NoSpecBuffer.d
The tst.NoSpecBuffer.d testcase exhibited some issues with timing
in relation to the trigger executable that was being used.  The
test does not actually require a trigger for the functionality that
is being tested, so it has been rewritten to no longer use a trigger
and use tick-* probes instead.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>