Prevent race between shutdown and process termination causing segfaults.
Even after the fix for bug 22824594, we still have a problem with
a concurrent dtrace shutdown and process termination.  The race is
as follows:

Main thread                       Control thread
(in dt_ps_proc_destroy())         (in dt_proc_control_cleanup())
                                  dt_proc_dpr_lock(dpr);
if (dpr->dpr_tid) {
	dt_proc_dpr_lock(dpr);
		(blocks)
		                  dt_proc_dpr_unlock(dpr);
				  dpr->dpr_tid = 0;
	...
	pthread_cancel(dpr->dpr_tid);

This now tries to cancel the thread with ID 0 and coredumps.

Even if you avoid this, there is *another* problem lurking in that
innocent-looking ..., because it's a proxy call to tell the
process-control thread to stop monitoring the state of the child:
calling that unlocks the dpr_lock (which we just locked).

Fixing both requires taking out the dpr_lock around the test of
dpr->dpr_tid instead of just after it, and calling proxy_monitor()
(which must be called under the dpr_lock) *before* testing the dpr_tid
(at this point, we don't care if the control thread is monitoring:
either the daemon thread is not dead, and we're cancelling it and
letting it clean up anyway, or it is dead, and we're Prelease()ing
ourselves.)

But that's not quite enough, because the control thread's handling of
the dpr_lock is also racy.  The problem is that it sets the dpr_tid
outside the dpr_lock: it has to, because dt_proc_dpr_unlock() needs the
value of dpr_tid to be unchanged to tell which lock count it should be
manipulating.  So add a dpr->dpr_ending flag, which is used only by the
cleanup code to ask the dt_proc_dpr_lock to zero out the dpr_tid once
it's tested it but before it unlocks the mutex (we can't add a
parameter, because the unlocking might be done directly or might be done
via a convoluted route involving Prelease() and the Ptrace() lock hook).

With both of these fixed, the bug appears to be resolved.

Also add a test for termination races of this nature, which uses a new
trigger which sleeps for the number of nanoseconds specified in the
delay environment variable.  The test itself is quite intricate, a state
machine which repeatedly fires off dtrace with a timeout probe and a
probe on syscall::exit_group:entry, then uses information on which of
these fired to dynamically hunt for the region between which dtrace
flips inconsistently between shutting down and spotting process
termination (the narrowness of this window means that this bug is only
likely to be evidenced on virtual machines, with their comparatively
higher variability of load), then walks up that region in ever-finer
slices, waiting for dtrace to dump core.

This takes, well, as long as you want it to take, but at least a few
hundred seconds to find the variable region, even when binary searching:
by default I've set it to 1500s, a completely arbitrary value.  As with
all the expensive tests, users are more or less expected to set the
runtest --timeout to as high a value as they can bear and go to sleep or
read War and Peace in the meantime.

Orabug: 23028026