Exit properly when all dtrace -c processes end, even if waitfd() failed.
libdtrace contains a process-death notifier which is called whenever a
monitored process terminates: dtrace(1) relies on this machinery to
decrement a count of running processes it started so that it can
terminate when all processes started via dtrace -c have exited.

Unfortunately, this notification is queued by the control thread cleanup
process, and is obviously not queued if a grab was noninvasive (because
in that case the control thread terminates immediately while the
monitored process is sitll running). Normally, such grabs are for things
like ustack(), where we don't care about the process death notification:
however, if waitfd() does not work, this happens for processes started
via -c and -p too, and we definitely do want to know when a -c process
dies.

So add a new function dt_proc_enqueue_exits() which dt_proc_sleep()
calls right before it scans the notification list; if anything has been
"demoted to noninvasiveness" due to a nonfunctional waitfd(), this scans
every running process for any running processes with no control thread,
and explicitly wait()s for them.  (Normally, nothing will have been so
demoted, so this doesn't impair performance in the normal case.)

In the process, we add extra machinery allowing dt_proc_notify() to be
called from a function that already holds the dph_lock (as
dt_proc_sleep() does), and allowing it to skip broadcast of the
notification (because dt_proc_sleep() would be the intended receiver
anyway, and is running in the same thread).