Fix endianness problems with SPARC 32-bit.
Big-endian platforms, unlike little-endian ones, must compensate for
variations in the size of scalars Pread() from victim processes both by
transferring fewer bytes but also by transferring them to a later place
in the target and by blanking out the start of the target.  We were
doing the former, but not the latter.

This effectively torpedoed even simple breakpoint dropping in such
processes.