Unskip a bunch of sparc64 tests.
When we got -c working, we forgot that there were a lot of tests that were
being individually skipped because of the lack of working -c or things
that it implies (working process-grabbing, working ustack() symbol
resolution, etc).

Unskip all of those, and other things (like the resize stress tests
and the openret syscall test) that work at least some of the time on
sparc now.