Correctly identify the software-singlestep breakpoint PC.
The code in Pwait_handle_waitpid() which endeavours to figure out which
breakpoint has been hit is making the assumption that if you get a breakpoint
signal and you know you're in the middle of handling a breakpoint that the
instruction pointer must be the one for that breakpoint.  This is a premature
optimization: on a software-singlestepping platform like SPARC we know we're in
the middle of handling the original breakpoint, but while working over the
temporary singlestepping breakpoint, our program counter has moved on to the
next instruction: we must handle *that* breakpoint, the temporary one, and not
assume this is another invocation of the original breakpoint.  (If we do, we end
up assuming that the breakpoint instruction we just poked in is in fact the
original instruction for the breakpoint, leading to recursive disaster the next
time the breakpoint is hit.)

This was never detected before because the only breakpoints we invoke more than
once are the ones on DT_DEBUG etc, and those were never fired by anything in the
testsuite until USDT was added and the tst.dlclose* tests started to be used.

On hardware singlestepping platforms, we must use the same mechanism as before:
observing the program counter after a hardware singlestep will put us somewhere
other than a breakpoint locus we know about, and quite possibly a very long way
away if the singlestepped instruction was a function return.

With this bug fixed, ld.so introspection works properly on SPARC, so we can fix
up a bunch of SPARC test results: the only difference from x86-64 pldd test
output is now the difference in the name of the 64-bit dynamic linker.

Orabug: 24454127