headers: eliminate need for arg counting in usdt macros
This uses the same machinery as the corresponding kernel change,
with adjustments for the distinct form of USDT probes (explicitly-
specified provider name, casts to unsigned long, etc).

Because the zero-argument case needs __VA_ARGS__ comma elision, a GNU C
extension which is not even in C99, we use the old macros unless
using the appropriate compiler and standard.  (C++ is always assumed
suitable on the grounds that the C++ headers require GNU C extensions
to be enabled: this may be inappropriate for non-GCC compilers, but it
*is* supported by LLVM for several years now.

Orabug: 24678905