libproc: Allow self-grabs.
We have long prohibited grabs of a running dtrace process by itself, on
the grounds that the first thing we do with such a process is stop it to
drop breakpoints, and that stopping yourself (or, indeed, ptrace()ing
yourself) is ridiculous and impossible to get out of.

However, now that we have noninvasive grabs which avoid all ptrace()
use, we can use that for dtrace self-grabs and get fairly reasonable
backtraces out (at least of libdtrace symbols).  This does mean we can't
drop any breakpoints, but they're useless in dtrace(1) in any case: we
never dlopen() so we don't need to refresh our symbol tables when that
happens, and we can't possibly grab ourself before main() so we don't
need -x evaltime support either.  (The process is also guaranteed
non-notifiable: we can't possibly have invoked ourselves via -c, and in
any case it's hard to notify us of our own death after we have died.)

This does mean that dtrace consumers can now monitor the operation of
their non-dtrace components, which was ruled out before, though
obviously they should look out because the dtrace components might well
cause a lot of spurious self-generated output if you monitor something
like the ioctl syscall.

Also add a test to make sure that if we do grab ourself, we don't stop
running, and we don't see a recurrence of bug 24829169 either.

Orabug: 24829169