tst.system-daemon.sh: Do not fail when glibc is upgraded
Under baremetal testing, the newly-installed system has a tendency to
run prelink and unlink libc.so while the test is running.  This then
leads to the system-daemon noninvasive tracing test being unable to get
at the running libc, losing access to symbols resolved from there.

This is not a failure but expected operation (there is no way to get
at the symbols of an unlinked binary without invasive tracing and
PTRACE_GETMAPFD, at least not yet).

Orabug: 22170826