dtrace: support Linux 4.3+ /proc/$pid/map_files.
Linux 4.3+ introduces a /proc/<pid>/map_files interface providing access
to the complete contents of all files mapped by all processes the user
can ptrace(), whether unlinked or not.  This is exactly the semantics
dtrace needs, and means we can drop our pre-existing PTRACE_GETMAPFD
local patch in favour of an upstreamed interface, while gaining extra
capabilities in the process; since /proc/<pid>/map_files only requires
the *ability* to ptrace() -- since you don't actually need to
ptrace-stop the process, you can use it even for processes you can't
afford to halt.  From the user's perspective, this means accurate symbol
resolution for all binaries on a Linux 4.3+ system, whether or not they
are system daemons, the dtrace executable itself or other things that it
is not safe to ptrace().  It requires CAP_SYS_ADMIN for now, but since
unprivileged dtracing does not exist yet this is no additional
restriction (and we can lift that constraint in the kernel, if need be,
since it adds no capabilities to the user that weren't already present
in PTRACE_GETMAPFD).

We already have a fallback for the cannot-ptrace case -- simply reading
the executable directly off disk if its (dev, ino) pair is the same as
the one in /proc/$pid/maps -- but this fallback cannot handle the case
of a binary being unlinked, as happens in prelink and upgrades.

The ptrace() interface allows us to simply pass in an address: ptrace()
figures out the right address range to look in, while the map_files
interface requires us to know the range ourselves (the files in the
map_files directory are named identically to the address-range string at
the start of /proc/$pid/maps entries).  We implement this by remembering
that string in a new field in the "struct prmap" that represents a
single executable mapping, then looking up the file at
Pbuild_file_symtab time, before we try to use PTRACE_GETMAPFD.  As an
optimization (and because we are likely to look up a lot of files in
this directory) we keep pathname lookups down by opening
/proc/$pid/map_files as an O_DIRECTORY and using openat().  An -ENOENT
failure causes us to never look up /proc/$pid/map_files again, on the
grounds that it doesn't exist in this kernel, so looking it up is a
waste of time.

As a minor additional feature, if PTRACE_GETMAPFD fails for whatever
reason we now fall back to trying to use open() on the disk file iff it
is the same (dev, ino) pair as the mapping in /proc/$pid/maps: this
can't really happen but it makes the code simpler. :)

The recent change to tst.system-daemon.sh to XFAIL if any mappings have
been deleted in the daemon we are testing (syslogd) can now be narrowed
a bit: the test will pass even if mappings are deleted, as long as
/proc/$pid/map_files exists.

Orabug: 24843582