Fix spurious not-a-memory-leak, and clean up pointless debuggging.
Valgrind complains about memory leaks whenever dtrace -C of a pipe
or stdin is used.  This is a false positive: it's not detecting a
leak from the main program, but complaining about the forked
child running the pipe-conversion splice() (see commit 9d66ec82b).
This exits inside dt_preproc() and is not a real leak, but we can
quiesce the warning and gain a tiny bit of efficiency by using
alloca() instead of malloc() (this is quite safe unless the user
specifies vast numbers of cpp args, and that's only doable via a
dtrace pragma, thus something that only the invoker can do: the only
negative consequence on overflow is a stack overrun and process
termination before any probes are enabled).

While we're at it, discard some useless debugging that gets the size of
the cpp output file (at the cost of a pointless seek, flush, and stat),
and quash an uninitialized warning that GCC cannot detect is spurious.