Make a start at reviving the ip tests.
These tests were entirely unported, relied on features of ping
that do not exist in the OL6/7 ping (notably UDP pinging), and
found remote hosts to work on by broadcast-pinging the local
subnet, which is not much use these days when no hosts respond.

Make everything look at the local gateway instead: it's easy to
identify and on the VM test hosts is probably the VM host itself,
easy to add stuff to.  (For all tests to be unskipped, it must
be running ssh and rpcbind on IPv4 and be answering both IPv4
and v6 pings.)

Untested, of course -- we don't have the ip provider yet.