tests: signals.sh: avoid race between dtrace startup and signal send
This test has multiple racy bits in it.  For now, we ignore the place
where we determine what signals do by default by starting test
processes, waiting a bit for them to start and then blasting them with
signals: in practice as long as the process is there at all, it'll work,
and it doesn't go wrong on any system yet observed.

But the second part, which kicks up a sleeping process, fires up
dtrace -p on it, then waits for a second for dtrace to start before
blasting the process with signals and observing dtrace's debugging
output to see if it's handling it right, *is* racy: it can easily
take dtrace longer than a second to start, and in that case we
blast the test process before dtrace has grabbed it, and the test
fails.  It's also inefficient: if dtrace starts much faster, we
wait even though we don't need to.

So, instead, exploit the fact that we can see the dtrace debugging
output by checking every .25s for a diagnostic debugging output
indicating that dtrace has completed its grab, and only then
blast the process with signals.  On a non-loaded system this speeds
up testing by a few seconds: on a heavily-loaded one, it should make
the tests work where before they had no hope.

Orabug: 24364338