libproc: do not flip state before calling Pbkpt_continue()
On SPARC, we see output in the debug log like

Unexpected ptrace (PTRACE_GETREGSET) error: Success

This is because getregs_sparc64() checks to see if the process is not in
stopped state before trying a getregs, since this is a ptrace operation
which requires the process to be stopped first.  Unfortunately, when
we do a Puntrace() and resume from stopped state, we reset the state
to the previous state popped from the stack (often PS_RUN) before
calling Pbkpt_continue() to resume -- but on some code paths
Pbkpt_continue() calls Pget_bkpt_ip(), which ultimately calls
getregs_sparc64() on SPARC.  The process is still actually stopped, but
because we reset our idea of its state too soon, we don't know this, so
spray this error message and end up concluding that the process is not
stopped, even though it is.

Fixing this is a little arcane: we cannot pop the process state before
calling Pbkpt_continue(), but because it often changes the state itself
(if stopped at a breakpoint, for instance), we cannot always do it
afterwards either.  So adjust Pbkpt_continue() to note whether it has
changed the process state in its return value, and have the call sites
that care about this check before adjusting the state again.

Orabug: 25711533 (but not a fix for this yet)