tests: filter the vDSO out of libproc create/grab pldd tests
These tests test libproc's phdr iteration and library search-path lookup
functions by running an enhanced version of ldd(1) on a test binary
(enhanced in that it can print the search path in force for each shared object,
which ldd cannot).  However, this also prints out the internal entry for the
vDSO, which is problematic because the vDSO is not present on all systems
(in particular, it is not present on upstream SPARC kernels without the UEK
patches installed).  The lack of the vDSO does not mean that libproc is
broken, so filter it out so its presence or absence doesn't affect us.