runtest: support test.x and test.*.x
These files are like other .x files: executable files echoing a reason
why a test is skipped or xfailed, and exiting 1 if it should xfail,
2 if it should be skipped.  However, like test.options they apply
to entire directories or supersets of directories.  If an .x file
and one or more test.options both exist, the "most faily" is taken:
skips supersede xfails supersede passes.

Orabug: 25815242
Thanks-to: Alan Maguire <alan.maguire@oracle.com> for the idea and
           initial implementation.