runtest: always log unpreprocessed test output
This means that runtest.log contains all addresses, etc, reported by
DTrace.  Test comparisons are still done on the postprocessed output,
and the diffs that are additionally logged if this goes wrong are
diffs of the postprocessed output, as before.

The .sum file contains postprocessed output, as it always has.