Support building against kernels built with O=
If you build your kernel with O=foo, the .config, object files and
generated headers end up in the subdirectory foo/.  The header-
inclusion machinery used to substitute D library files does not
understand this, and fails on such kernels.

Worse, the include order is not the same as the one the kernel uses,
possibly leading to inaccurate substitution. The order seen in this
commit is derived directly from inspection of the kernel makefiles, and
seems to work fine.

To build DTrace against a locally-built kernel built with O=foo, pass
KERNELODIR=foo to the DTrace build (the same value).