Support C-based tests for DTrace consumers.
After this change, you can drop a .c file into a test directory and have
it linked with -ldtrace automatically.  Headers and libdtrace itself are
picked up from the build tree, unless this is an installed DTrace
testsuite package.  You can override the libraries to link with with the
new @@link option in a comment in the C test program. (There is as yet
no support for overriding CPPFLAGS, CFLAGS, or LDFLAGS on a test-by-test
basis, but it would be easy to add if needed.)

Triggers specified via @@trigger or a .t file should work, as should
other features like xfails, results comparison, results postprocessors,
etc.  There is no analogue of no-execution tests (.d without
corresponding .r): the program, if successfully compiled, is always run.