runtest: stop lying about whether we executed a test or not
The dtrace consumer test revamp worsed a pre-existing bug where D tst.*
and err.* tests without trigger statements could be considered in the
output messages not to have executed, even though all such tests are
always executed (syntax checks don't even make sense for them).

Rather than assuming that tests are noexec if they don't have a trigger,
or that all .d tests are noexec (huh?), move the check for "have we
got a trigger, are we noexec, do we tell the caller we're noexec" into
the place that runs .d tests, so they never get out of sync.