Pread(): sparc: fix spurious errors and misreading of unaligned data
The code to read data from high addresses (needed on sparc, but not on
x86) fails to adjust the length to read to account for the fact that
only whole numbers of sizeof(unsigned long) units can be read, or
to read from the right place, or to set the number of bytes properly
afterwards.

The result is that unaligned reads (as done by e.g. the glibc
l_searchlist searching code) read the wrong data and claim that they
read more bytes than they were asked to.

Orabug: 26378141