tests/unittest/tcp: Only TCP tests using client.ip.pl need IO::Socket::IP
It turns out that some but not all the TCP tests do fail in the absence
of IO::Socket::IP; specifically, those tests that use ../ip/client.ip.pl
will fail in that situation.

So check for the actual failure case, by grepping the test we are
running in the .x script.

Orabug: 26419357